code-smell-generation-assistant/requirements.txt
python-dotenv==0.20.0
openai
